email-consumer/requirements.txt
pika==1.3.2
sendgrid==6.11.0
python-dotenv==1.0.1
